+++ tools/BugReportTool/BugReportTool/BugReportTool.vcxproj
@@ -38,7 +38,7 @@
     </ClCompile>
     <ClCompile Include="ReportMonitorInfo.cpp" />
     <ClCompile Include="Main.cpp" />
-    <ClCompile Include="ReportRegistry.cpp" />
+    <ClCompile Include="RegistryUtils.cpp" />
     <ClCompile Include="ZipTools\zipfolder.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -54,7 +54,7 @@
     <ClInclude Include="..\..\..\deps\cziplib\src\zip.h" />
     <ClInclude Include="ReportMonitorInfo.h" />
     <ClInclude Include="..\..\..\common\utils\json.h" />
-    <ClInclude Include="ReportRegistry.h" />
+    <ClInclude Include="RegistryUtils.h" />
     <ClInclude Include="ZipTools\zipfolder.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
