+++ tools/BugReportTool/BugReportTool/BugReportTool.vcxproj.filters
@@ -10,7 +10,7 @@
     </ClCompile>
     <ClCompile Include="..\..\..\deps\cziplib\src\zip.c" />
     <ClCompile Include="ReportMonitorInfo.cpp" />
-    <ClCompile Include="ReportRegistry.cpp" />
+    <ClCompile Include="RegistryUtils.cpp" />
   </ItemGroup>
   <ItemGroup>
     <Filter Include="ZipTools">
@@ -25,6 +25,6 @@
     <ClInclude Include="..\..\..\deps\cziplib\src\miniz.h" />
     <ClInclude Include="..\..\..\deps\cziplib\src\zip.h" />
     <ClInclude Include="ReportMonitorInfo.h" />
-    <ClInclude Include="ReportRegistry.h" />
+    <ClInclude Include="RegistryUtils.h" />
   </ItemGroup>
 </Project>
