+++ tools/BugReportTool/BugReportTool/RegistryUtils.h
